+++ test/main/java/edu/wpi/rail/jrosbridge/messages/geometry/TestPose.java
@@ -171,7 +171,7 @@ public class TestPose {
 	}
 
 	@Test
-	public void testSetType() {
+	public void testSetMessageType() {
 		empty.setMessageType("test");
 		assertEquals("test", empty.getMessageType());
 	}