+++ test/main/java/edu/wpi/rail/jrosbridge/messages/geometry/TestPoseWithCovariance.java
@@ -102,7 +102,7 @@ public class TestPoseWithCovariance {
 	// }
 
 	@Test
-	public void testSetType() {
+	public void testSetMessageType() {
 		empty.setMessageType("test");
 		assertEquals("test", empty.getMessageType());
 	}