+++ test/main/java/edu/wpi/rail/jrosbridge/messages/geometry/TestTransform.java
@@ -185,7 +185,7 @@ public class TestTransform {
 	}
 
 	@Test
-	public void testSetType() {
+	public void testSetMessageType() {
 		empty.setMessageType("test");
 		assertEquals("test", empty.getMessageType());
 	}